Provide a more typical interface for exponential powers: an exp function.
Note that this does take away 'e' from the Prelude interface, but prelude.js
previously implemented it as 'Math.e', which is not a valid identifier (Math.E
is defined according to ECMA-262), so it is unlikely to break any code that
previously worked. Code that needs exponential e can now use (exp 0), and
the more common case, evaluating powers of e, is usually more efficient using
the specialised Math.exp function than it is to use a general power function.